tests fixes

+++ app/Http/Controllers/CompanySettingsLogoController.php
@@ -13,7 +13,7 @@ class CompanySettingsLogoController extends Controller{
 
 	public function show(Request $request){
 
-		$company_id = Sanitize::input($request->input('company_id'));
+		$company_id = (int) Sanitize::input($request->input('company_id'));
 		$path = 'logos/'.$company_id;
 
 		$company = General::fetchDefaultCompanyById($company_id);
